Grok2_version/requirements.txt
# Blockchain Interaction
web3==6.15.1
python-binance==1.0.19
solana==0.30.2

# AI/ML Libraries
scikit-learn==1.4.1.post1
tensorflow==2.15.0
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1

# GUI Development
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

# Data Processing & Visualization
ccxt==4.2.15
matplotlib==3.8.3

# Add testing dependencies
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-mock==3.12.0
coverage==7.4.3 